Primeros, Repository y Resource
-- Departamento
--provincia
--distrito
--area
--institucion

+++ src/main/java/com/clinica/erp/model/DetalleVentaKey.java
@@ -6,8 +6,8 @@ import javax.persistence.Column;
 
 public class DetalleVentaKey  implements Serializable  {
     @Column(name = "idVenta")
-    Long idVenta;
+    Integer idVenta;
 
     @Column(name = "idServicio")
-    Long idServicio;
+    Integer idServicio;
 }
