+++ src/main/java/com/clinica/erp/repository/DepartamentoRepository.java
@@ -4,6 +4,6 @@ import com.clinica.erp.model.Departamento;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartamentoRepository extends JpaRepository<Departamento, Long>{
+public interface DepartamentoRepository extends JpaRepository<Departamento, Integer>{
     
 }
